refactor: replace heavy SVG token icons with PNGs
