+++ .changeset/mean-dogs-appear.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": patch
+---
+
+replace heavy SVG token icons with PNGs
